fix calculation of remaining buffer size when calling update scripts
the buffer used to hold the full shell command to call the upgrade
scripts during FOTA is of length ARM_UC_MAXIMUM_COMMAND_LENGTH
(currently 256), but the caluclation for space remaining after printing
a file path to the buffer was incorrectly using a different value,
ARM_UC_MAXIMUM_FILE_AND_PATH_LENGTH (currently 128) as the buffer
length, causing the function to fail incorrectly.

the fix is to use the buffer length passed into the function rather
than using a different hardcoded #define.