Add network proxy support
* add CLI option "--proxy" to set a network proxy config
* add CMake options to turn on network proxy support

This commit also contains a small change to workaround
a mbed-cloud-client API memory issue. The mbed-cloud-client
API expects the caller to maintain the memory buffer that supplies
the proxy address, auth, and creds, and so the local variables
used to pass the proxy configs to the mbed-cloud-client library
had to be declared static so that they are not destroyed after
the config function ends.