Make tests run again
The curl include was not found, as it was not specified to be included.
This makes the tests run again - 160/160 and 8/8.

Not guaranteeing the results, though - lots of issues in the logs
but some of those might be really negative tests cases?

OK (160 tests, 160 ran, 18293 checks, 0 ignored, 0 filtered out, 169 ms)

[ERR ][edge-io]: Cannot remove the socket lock file: /tmp/foo.lock
[ERR ][edge-io]: Cannot unlock the socket lock: /tmp/foo.lock
[ERR ][edge-io]: The socket lock '/tmp/foo.lock' is held by another process. Is there another Edge Core running?
[ERR ][edge-io]: The socket lock file '/tmp/foo.lock' cannot be created. Please check the permissions!
...
OK (8 tests, 8 ran, 31 checks, 0 ignored, 0 filtered out, 1 ms)