Fix failing test cases broken by log changes
Fix the unit tests that are failing due to logs being modified.
The tests are using STRCMP to check if the failure case is matching
the test case and since more details were added to the functions
(i.e. failing function name) -> the strings are not matching anymore.

Fix those. Very easy for the most cases, but the test on line 5152
is more tricky, as it seems to be used for two different failure
cases and those are failing in different functions.
- So, break that test into steps which looks for pieces of the strings
  rather than verifying the whole string.